Rogue like dinosaur park game
This is a team project focusing on object oriented programming and design principles. This is a relatively large project, so expect a lot of files in the zip file. You can simply run the jar file in the command prompt to test it out. Note that the focus was not game design, hence no fancy graphics.